EQ-4-174-Changed (hopefully) all service calls that return offers to account for the session

+++ audace-backend/src/main/java/com/equipe4/audace/service/EmployerService.java
@@ -4,9 +4,13 @@ import com.equipe4.audace.dto.EmployerDTO;
 import com.equipe4.audace.dto.offer.OfferDTO;
 import com.equipe4.audace.model.Employer;
 import com.equipe4.audace.model.offer.Offer;
+import com.equipe4.audace.model.offer.OfferSession;
+import com.equipe4.audace.model.session.Session;
 import com.equipe4.audace.repository.EmployerRepository;
 import com.equipe4.audace.repository.offer.OfferRepository;
+import com.equipe4.audace.repository.offer.OfferSessionRepository;
 import com.equipe4.audace.repository.security.SaltRepository;
+import com.equipe4.audace.utils.SessionManipulator;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
@@ -17,15 +21,21 @@ import java.util.Optional;
 public class EmployerService extends GenericUserService<Employer> {
     private final EmployerRepository employerRepository;
     private final OfferRepository offerRepository;
+    private final OfferSessionRepository offerSessionRepository;
+    private final SessionManipulator sessionManipulator;
 
     public EmployerService(
             SaltRepository saltRepository,
             EmployerRepository employerRepository,
-            OfferRepository offerRepository
+            OfferRepository offerRepository,
+            OfferSessionRepository offerSessionRepository,
+            SessionManipulator sessionManipulator
     ) {
         super(saltRepository);
         this.employerRepository = employerRepository;
         this.offerRepository = offerRepository;
+        this.offerSessionRepository = offerSessionRepository;
+        this.sessionManipulator = sessionManipulator;
     }
 
     @Transactional
@@ -45,31 +55,49 @@ public class EmployerService extends GenericUserService<Employer> {
         return employerRepository.findById(employerId).map(Employer::toDTO);
     }
 
-    public List<EmployerDTO> findAllEmployers(){
-        return employerRepository.findAll().stream().map(Employer::toDTO).toList();
-    }
-
+    @Transactional
     public Optional<OfferDTO> createOffer(OfferDTO offerDTO){
         if(offerDTO == null) throw new IllegalArgumentException("Offer cannot be null");
 
-        Offer offer = offerDTO.fromDTO();
-        offer.setOfferStatus(Offer.OfferStatus.PENDING);
+        Session session = sessionManipulator.getCurrentSession();
+        Offer offer = offerRepository.save(offerDTO.fromDTO());
 
-        return Optional.of(offerRepository.save(offer).toDTO());
+        offerSessionRepository.save(new OfferSession(null, offer, session));
+
+        return Optional.of(offer.toDTO());
     }
 
+    @Transactional
     public List<OfferDTO> findAllOffersByEmployerId(Long employerId){
         Employer employer = employerRepository.findById(employerId).orElseThrow();
-        return offerRepository.findAllByEmployer(employer).stream().map(Offer::toDTO).toList();
+        List<Offer> offers = offerRepository.findAllByEmployer(employer);
+
+        return sessionManipulator
+                .removeOffersNotInCurrentSession(offers)
+                .stream()
+                .map(Offer::toDTO)
+                .toList();
     }
 
+    @Transactional
     public Optional<OfferDTO> updateOffer(OfferDTO offerDTO){
         Offer offer = offerRepository.findById(offerDTO.getId()).orElseThrow();
+
+        if (!sessionManipulator.verifyIfOfferIsInCurrentSession(offer)) {
+            throw new IllegalStateException("Offer is not in current session");
+        }
+
         return Optional.of(offerRepository.save(offer).toDTO());
     }
 
+    @Transactional
     public void deleteOffer(Long offerId){
         Offer offer = offerRepository.findById(offerId).orElseThrow();
+
+        if (!sessionManipulator.verifyIfOfferIsInCurrentSession(offer)) {
+            throw new IllegalStateException("Offer is not in current session");
+        }
+
         offerRepository.delete(offer);
     }
 }
